builder-utils: Add a function to check appstream version requirements
This could've been done with the libappstream API but since
flatpak-builder has until now avoided linking to it and linking to it
now does not make sense just to check the version. That is postponed
until appstream-compose is API stable and we can port to it.

The CLI output format has been stable for 8+ years, see [1] [2]. We
prefer the library version in case of a mismatch.

[1]: https://github.com/ximion/appstream/commit/d83174a4695e13ef41730c7a055b7e9b9077c211
[2]: https://github.com/ximion/appstream/blob/56ecd0067495213a64f3a490cbbc030fcfd5b8f1/tools/appstreamcli.c#L1594-L1604