builder-manifest: Add argument to run compose with selected url policy
appstream-glib always composed with full urls. The URL policy got
unintentionallly switched to partial urls during libappstream port in
1.3.x.

Such a break in behaviour makes sense for unstable releases but we
cannot revert back to full urls in a stable release. Using full URLs
is desired for Flathub.

So add a cli arg to control the behaviour and default to partial url.
This may be flipped later in the next unstable release to default
to full urls again.

This is an alternative to https://github.com/flatpak/flatpak-builder/pull/576